Rename sample data - seed data
