+++ src/api/common/helpers/seed-database.test.js
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose'
 import { seedDatabase } from '~/src/api/common/helpers/seed-database.js'
-import data from '~/src/api/common/helpers/sample-data/index.js'
+import data from '~/src/api/common/helpers/seed-data/index.js'
 import models from '~/src/api/common/models/index.js'
 
 jest.mock('mongoose', () => ({
